facturation adapted, coupon added missing column, plan changed endpoint

+++ src/main/java/com/talenthub/AccountManager/repository/FacturationRepository.java
@@ -4,6 +4,6 @@ import com.talenthub.AccountManager.model.Facturation;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.JpaSpecificationExecutor;
 
-public interface FacturationRepository extends JpaRepository<Facturation, Integer>, JpaSpecificationExecutor<Facturation> {
+public interface FacturationRepository extends JpaRepository<Facturation, Long>, JpaSpecificationExecutor<Facturation> {
 
 }
